feat: Support pools without oracle
Various refactoring to support different pool types and moving stuff
around.

Included a e2e test for each of the alternative pools.

Fixed bug in test utils with assigning false return in lodash forEach.
