+++ src/dex/platypus/platypus-e2e.test.ts
@@ -23,8 +23,8 @@ describe('Platypus E2E', () => {
     const tokenASymbol: string = 'USDC';
     const tokenBSymbol: string = 'DAIE';
 
-    const tokenAAmount: string = '99999000000';
-    const tokenBAmount: string = '99999000000000000000000';
+    const tokenAAmount: string = '9999000000';
+    const tokenBAmount: string = '9999000000000000000000';
 
     const sideToContractMethods = new Map([
       [
@@ -36,6 +36,15 @@ describe('Platypus E2E', () => {
         ],
       ],
     ]);
+
+    const poolTokenSymbols: string[] = [
+      'newFRAX',
+      'MIM',
+      'YUSD',
+      'H2O',
+      'MONEY',
+      'TSD',
+    ];
 
     sideToContractMethods.forEach((contractMethods, side) =>
       contractMethods.forEach((contractMethod: ContractMethod) => {
@@ -66,6 +75,26 @@ describe('Platypus E2E', () => {
               provider,
             );
           });
+          if (
+            contractMethod === ContractMethod.simpleSwap &&
+            side === SwapSide.SELL
+          ) {
+            poolTokenSymbols.forEach(poolTokenSymbol => {
+              it(`${tokenASymbol} -> ${poolTokenSymbol}`, async () => {
+                await testE2E(
+                  tokens[tokenASymbol],
+                  tokens[poolTokenSymbol],
+                  holders[tokenASymbol],
+                  tokenAAmount,
+                  side,
+                  dexKey,
+                  contractMethod,
+                  network,
+                  provider,
+                );
+              });
+            });
+          }
         });
       }),
     );
