+++ src/dex/platypus/platypus.ts
@@ -16,102 +16,22 @@ import { IDex } from '../../dex/idex';
 import { IDexHelper } from '../../dex-helper/idex-helper';
 import {
   DexParams,
+  PlatypusOracleType,
   PlatypusData,
   PlatypusConfigInfo,
   PlatypusPoolState,
+  PlatypusPurePoolState,
 } from './types';
 import { SimpleExchange } from '../simple-exchange';
 import { PlatypusConfig, Adapters } from './config';
 import { ChainLinkSubscriber } from '../../lib/chainlink';
+import { PlatypusPoolBase } from './pool-base';
 import { PlatypusPool } from './pool';
-import { BI_POWS } from '../../bigint-constants';
+import { PlatypusPurePool } from './pool-pure';
 import ERC20ABI from '../../abi/erc20.json';
 import PoolABI from '../../abi/platypus/pool.json';
 import AssetABI from '../../abi/platypus/asset.json';
 import OracleABI from '../../abi/platypus/oracle.json';
-
-const ETH_UNIT = BI_POWS[18];
-const WAD = BI_POWS[18];
-const RAY = BI_POWS[27];
-
-function wmul(x: bigint, y: bigint): bigint {
-  return (x * y + WAD / 2n) / WAD;
-}
-
-function wdiv(x: bigint, y: bigint): bigint {
-  return (x * WAD + y / 2n) / y;
-}
-
-function rmul(x: bigint, y: bigint): bigint {
-  return (x * y + RAY / 2n) / RAY;
-}
-
-function rpow(x: bigint, n: bigint): bigint {
-  let z = n % 2n !== 0n ? x : RAY;
-
-  for (n /= 2n; n !== 0n; n /= 2n) {
-    x = rmul(x, x);
-
-    if (n % 2n !== 0n) {
-      z = rmul(z, x);
-    }
-  }
-
-  return z;
-}
-
-function slippageFunc(
-  k: bigint,
-  n: bigint,
-  c1: bigint,
-  xThreshold: bigint,
-  x: bigint,
-): bigint {
-  if (x < xThreshold) {
-    return c1 - x;
-  } else {
-    return wdiv(k, (rpow((x * RAY) / WAD, n) * WAD) / RAY);
-  }
-}
-
-function calcSlippage(
-  k: bigint,
-  n: bigint,
-  c1: bigint,
-  xThreshold: bigint,
-  cash: bigint,
-  liability: bigint,
-  cashChange: bigint,
-  addCash: boolean,
-): bigint {
-  const covBefore = wdiv(cash, liability);
-  let covAfter: bigint;
-  if (addCash) {
-    covAfter = wdiv(cash + cashChange, liability);
-  } else {
-    covAfter = wdiv(cash - cashChange, liability);
-  }
-  if (covBefore === covAfter) {
-    return 0n;
-  }
-
-  const slippageBefore = slippageFunc(k, n, c1, xThreshold, covBefore);
-  const slippageAfter = slippageFunc(k, n, c1, xThreshold, covAfter);
-
-  if (covBefore > covAfter) {
-    return wdiv(slippageAfter - slippageBefore, covBefore - covAfter);
-  } else {
-    return wdiv(slippageBefore - slippageAfter, covAfter - covBefore);
-  }
-}
-
-function calcSwappingSlippage(si: bigint, sj: bigint): bigint {
-  return WAD + si - sj;
-}
-
-function calcHaircut(amount: bigint, rate: bigint): bigint {
-  return wmul(amount, rate);
-}
 
 export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
   static readonly erc20Interface = new Interface(ERC20ABI);
@@ -123,7 +43,9 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
   protected cfgInfo?: PlatypusConfigInfo;
   protected poolLiquidityUSD?: { [poolAddress: string]: number };
 
-  protected eventPools?: { [poolAddress: string]: PlatypusPool };
+  protected eventPools?: {
+    [poolAddress: string]: PlatypusPool | PlatypusPurePool;
+  };
 
   readonly hasConstantPriceLargeAmounts = false;
 
@@ -155,15 +77,35 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
         (await this.dexHelper.web3Provider.eth.getCode(p.address, blockNumber))
           .length > 3
       ) {
-        cfgInfo.poolAddresses.push(p.address.toLowerCase());
+        const poolAddress = p.address.toLowerCase();
+        cfgInfo.poolAddresses.push(poolAddress);
+        if (p.oracleType === PlatypusOracleType.None) {
+          cfgInfo.pools[poolAddress] = {
+            oracleType: PlatypusOracleType.None,
+            tokenAddresses: [],
+            tokens: {},
+          };
+        } else if (p.oracleType === PlatypusOracleType.ChainLink) {
+          cfgInfo.pools[poolAddress] = {
+            oracleType: PlatypusOracleType.ChainLink,
+            priceOracleAddress: '',
+            tokenAddresses: [],
+            tokens: {},
+          };
+        } else {
+          throw new Error('Invalid pool oracle type in Platypus');
+        }
       }
     }
     let inputs: MultiCallInput[] = [];
     for (const poolAddress of cfgInfo.poolAddresses) {
-      inputs.push({
-        target: poolAddress,
-        callData: Platypus.poolInterface.encodeFunctionData('getPriceOracle'),
-      });
+      const pool = cfgInfo.pools[poolAddress];
+      if (pool.oracleType !== PlatypusOracleType.None) {
+        inputs.push({
+          target: poolAddress,
+          callData: Platypus.poolInterface.encodeFunctionData('getPriceOracle'),
+        });
+      }
       inputs.push({
         target: poolAddress,
         callData:
@@ -177,21 +119,20 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
     ).returnData;
     let i = 0;
     for (const poolAddress of cfgInfo.poolAddresses) {
-      const priceOracleAddress = Platypus.poolInterface
-        .decodeFunctionResult('getPriceOracle', returnData[i++])[0]
-        .toLowerCase();
-      const tokenAddresses = Platypus.poolInterface
+      const pool = cfgInfo.pools[poolAddress];
+      if (pool.oracleType !== PlatypusOracleType.None) {
+        pool.priceOracleAddress = Platypus.poolInterface
+          .decodeFunctionResult('getPriceOracle', returnData[i++])[0]
+          .toLowerCase();
+      }
+      pool.tokenAddresses = Platypus.poolInterface
         .decodeFunctionResult('getTokenAddresses', returnData[i++])[0]
         .map((a: Address) => a.toLowerCase());
-      cfgInfo.pools[poolAddress] = {
-        priceOracleAddress,
-        tokenAddresses,
-        tokens: {},
-      };
     }
     inputs = [];
     for (const poolAddress of cfgInfo.poolAddresses) {
-      for (const tokenAddress of cfgInfo.pools[poolAddress].tokenAddresses) {
+      const pool = cfgInfo.pools[poolAddress];
+      for (const tokenAddress of pool.tokenAddresses) {
         inputs.push({
           target: tokenAddress,
           callData: Platypus.erc20Interface.encodeFunctionData('symbol'),
@@ -206,13 +147,15 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
             tokenAddress,
           ]),
         });
-        inputs.push({
-          target: cfgInfo.pools[poolAddress].priceOracleAddress,
-          callData: Platypus.oracleInterface.encodeFunctionData(
-            'getSourceOfAsset',
-            [tokenAddress],
-          ),
-        });
+        if (pool.oracleType === PlatypusOracleType.ChainLink) {
+          inputs.push({
+            target: pool.priceOracleAddress,
+            callData: Platypus.oracleInterface.encodeFunctionData(
+              'getSourceOfAsset',
+              [tokenAddress],
+            ),
+          });
+        }
       }
     }
     returnData = (
@@ -222,7 +165,8 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
     ).returnData;
     i = 0;
     for (const poolAddress of cfgInfo.poolAddresses) {
-      for (const tokenAddress of cfgInfo.pools[poolAddress].tokenAddresses) {
+      const pool = cfgInfo.pools[poolAddress];
+      for (const tokenAddress of pool.tokenAddresses) {
         const tokenSymbol = Platypus.erc20Interface.decodeFunctionResult(
           'symbol',
           returnData[i++],
@@ -234,29 +178,39 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
         const assetAddress = Platypus.poolInterface
           .decodeFunctionResult('assetOf', returnData[i++])[0]
           .toLowerCase();
-        const proxyAddress = Platypus.oracleInterface
-          .decodeFunctionResult('getSourceOfAsset', returnData[i++])[0]
-          .toLowerCase();
-        cfgInfo.pools[poolAddress].tokens[tokenAddress] = {
-          tokenSymbol,
-          tokenDecimals,
-          assetAddress,
-          chainlink: {
-            proxyAddress,
-            aggregatorAddress: '',
-          },
-        };
+        if (pool.oracleType === PlatypusOracleType.ChainLink) {
+          const proxyAddress = Platypus.oracleInterface
+            .decodeFunctionResult('getSourceOfAsset', returnData[i++])[0]
+            .toLowerCase();
+          pool.tokens[tokenAddress] = {
+            tokenSymbol,
+            tokenDecimals,
+            assetAddress,
+            chainlink: {
+              proxyAddress,
+              aggregatorAddress: '',
+            },
+          };
+        } else {
+          pool.tokens[tokenAddress] = {
+            tokenSymbol,
+            tokenDecimals,
+            assetAddress,
+          };
+        }
       }
     }
     inputs = [];
     for (const poolAddress of cfgInfo.poolAddresses) {
-      for (const tokenAddress of cfgInfo.pools[poolAddress].tokenAddresses) {
-        inputs.push(
-          ChainLinkSubscriber.getReadAggregatorMultiCallInput(
-            cfgInfo.pools[poolAddress].tokens[tokenAddress].chainlink
-              .proxyAddress,
-          ),
-        );
+      const pool = cfgInfo.pools[poolAddress];
+      if (pool.oracleType === PlatypusOracleType.ChainLink) {
+        for (const tokenAddress of pool.tokenAddresses) {
+          inputs.push(
+            ChainLinkSubscriber.getReadAggregatorMultiCallInput(
+              pool.tokens[tokenAddress].chainlink.proxyAddress,
+            ),
+          );
+        }
       }
     }
     returnData = (
@@ -266,12 +220,12 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
     ).returnData;
     i = 0;
     for (const poolAddress of cfgInfo.poolAddresses) {
-      for (const tokenAddress of cfgInfo.pools[poolAddress].tokenAddresses) {
-        cfgInfo.pools[poolAddress].tokens[
-          tokenAddress
-        ].chainlink.aggregatorAddress = ChainLinkSubscriber.readAggregator(
-          returnData[i++],
-        ).toLowerCase();
+      const pool = cfgInfo.pools[poolAddress];
+      if (pool.oracleType === PlatypusOracleType.ChainLink) {
+        for (const tokenAddress of pool.tokenAddresses) {
+          pool.tokens[tokenAddress].chainlink.aggregatorAddress =
+            ChainLinkSubscriber.readAggregator(returnData[i++]).toLowerCase();
+        }
       }
     }
     return cfgInfo;
@@ -292,20 +246,40 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
       throw new Error(
         'initializePricing: Platypus cfgInfo still null after init',
       );
-    const eventPools: { [poolAddress: string]: PlatypusPool } = {};
+    const eventPools: {
+      [poolAddress: string]: PlatypusPool | PlatypusPurePool;
+    } = {};
     for (const poolAddress of this.cfgInfo.poolAddresses) {
-      const pool = new PlatypusPool(
-        this.dexKey,
-        this.network,
-        this.config.pools.find(
-          p => p.address.toLowerCase() === poolAddress,
-        )!.name,
-        poolAddress,
-        this.cfgInfo.pools[poolAddress],
-        this.dexHelper,
-      );
-      const state = await pool.generateState(blockNumber);
-      pool.setState(state, blockNumber);
+      const cfgPool = this.cfgInfo.pools[poolAddress];
+      const poolName = this.config.pools.find(
+        p => p.address.toLowerCase() === poolAddress,
+      )!.name;
+      let pool: PlatypusPool | PlatypusPurePool;
+      if (cfgPool.oracleType === PlatypusOracleType.None) {
+        pool = new PlatypusPurePool(
+          this.dexKey,
+          this.network,
+          poolName,
+          poolAddress,
+          cfgPool,
+          this.dexHelper,
+        );
+      } else if (cfgPool.oracleType === PlatypusOracleType.ChainLink) {
+        pool = new PlatypusPool(
+          this.dexKey,
+          this.network,
+          poolName,
+          poolAddress,
+          cfgPool,
+          this.dexHelper,
+        );
+      } else {
+        throw new Error('Platypus cfgInfo invalid pool type');
+      }
+      await (async <P>(p: P extends PlatypusPoolBase<infer S> ? P : never) => {
+        const state = await p.generateState(blockNumber);
+        p.setState(state, blockNumber);
+      })(pool);
       this.dexHelper.blockManager.subscribeToLogs(
         pool,
         pool.addressesSubscribed,
@@ -357,63 +331,6 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
     ).map(p => this.getPoolIdentifier(p));
   }
 
-  protected computePrices(
-    srcTokenAddress: Address,
-    srcTokenDecimals: number,
-    destTokenAddress: Address,
-    destTokenDecimals: number,
-    amounts: bigint[],
-    state: PlatypusPoolState,
-  ): bigint[] {
-    const tokenAPrice = state.chainlink[srcTokenAddress].answer;
-    const tokenBPrice = state.chainlink[destTokenAddress].answer;
-    if (tokenBPrice > tokenAPrice) {
-      if (
-        ((tokenBPrice - tokenAPrice) * ETH_UNIT) / tokenBPrice >
-        state.params.maxPriceDeviation
-      ) {
-        return Array(amounts.length).fill(0n);
-      }
-    } else {
-      if (
-        ((tokenAPrice - tokenBPrice) * ETH_UNIT) / tokenAPrice >
-        state.params.maxPriceDeviation
-      ) {
-        return Array(amounts.length).fill(0n);
-      }
-    }
-    return amounts.map(fromAmount => {
-      const idealToAmount =
-        (fromAmount * getBigIntPow(destTokenDecimals)) /
-        getBigIntPow(srcTokenDecimals);
-      if (state.asset[destTokenAddress].cash < idealToAmount) return 0n;
-      const slippageFrom = calcSlippage(
-        state.params.slippageParamK,
-        state.params.slippageParamN,
-        state.params.c1,
-        state.params.xThreshold,
-        state.asset[srcTokenAddress].cash,
-        state.asset[srcTokenAddress].liability,
-        fromAmount,
-        true,
-      );
-      const slippageTo = calcSlippage(
-        state.params.slippageParamK,
-        state.params.slippageParamN,
-        state.params.c1,
-        state.params.xThreshold,
-        state.asset[destTokenAddress].cash,
-        state.asset[destTokenAddress].liability,
-        idealToAmount,
-        false,
-      );
-      const swappingSlippage = calcSwappingSlippage(slippageFrom, slippageTo);
-      const toAmount = wmul(idealToAmount, swappingSlippage);
-      const haircut = calcHaircut(toAmount, state.params.haircutRate);
-      return toAmount - haircut;
-    });
-  }
-
   // Returns pool prices for amounts.
   // If limitPools is defined only pools in limitPools
   // should be used. If limitPools is undefined then
@@ -445,22 +362,28 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
               limitPools.includes(this.getPoolIdentifier(poolAddress)),
           )
           .map(poolAddress =>
-            (async () => {
-              let state = this.eventPools![poolAddress].getState(blockNumber);
+            (async <P>(
+              pool: P extends PlatypusPoolBase<infer S> ? P : never,
+            ) => {
+              let state = pool.getState(blockNumber);
               if (!state) {
-                state = await this.eventPools![poolAddress].generateState(
-                  blockNumber,
-                );
-                this.eventPools![poolAddress].setState(state, blockNumber);
+                state = await pool.generateState(blockNumber);
+                pool.setState(state, blockNumber);
               }
               if (state.params.paused) return null;
-              const [unit, ...prices] = this.computePrices(
-                srcTokenAddress,
-                this.cfgInfo!.pools[poolAddress].tokens[srcTokenAddress]
-                  .tokenDecimals,
-                destTokenAddress,
-                this.cfgInfo!.pools[poolAddress].tokens[destTokenAddress]
-                  .tokenDecimals,
+              const [unit, ...prices] = pool.computePrices(
+                {
+                  address: srcTokenAddress,
+                  decimals:
+                    this.cfgInfo!.pools[poolAddress].tokens[srcTokenAddress]
+                      .tokenDecimals,
+                },
+                {
+                  address: destTokenAddress,
+                  decimals:
+                    this.cfgInfo!.pools[poolAddress].tokens[destTokenAddress]
+                      .tokenDecimals,
+                },
                 [getBigIntPow(srcToken.decimals), ...amounts],
                 state,
               );
@@ -476,7 +399,7 @@ export class Platypus extends SimpleExchange implements IDex<PlatypusData> {
                 poolIdentifier: this.getPoolIdentifier(poolAddress),
               };
               return ret;
-            })(),
+            })(this.eventPools![poolAddress]),
           ),
       )
     ).filter((p): p is PoolPrices<PlatypusData> => !!p);
