+++ tests/utils-events.ts
@@ -113,15 +113,13 @@ export function deepTypecast<T>(
   checker: (val: any) => boolean,
   caster: (val: T) => any,
 ): any {
-  return _.forEach(
-    obj,
-    (val: any, key: any, obj: any) =>
-      (obj[key] = checker(val)
-        ? caster(val)
-        : _.isObject(val)
-        ? deepTypecast(val, checker, caster)
-        : val),
-  );
+  return _.forEach(obj, (val: any, key: any, obj: any) => {
+    obj[key] = checker(val)
+      ? caster(val)
+      : _.isObject(val)
+      ? deepTypecast(val, checker, caster)
+      : val;
+  });
 }
 
 export function getSavedConfig<Config>(
